% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions_kinetics_DE.R
\name{combine_lineages_COSICC_DE}
\alias{combine_lineages_COSICC_DE}
\title{combine_lineages_COSICC_DE}
\usage{
combine_lineages_COSICC_DE(
  list_of_COSICC_DE_core_outputs,
  names_lineage_trajectories,
  thresh_p_adj = 0.1
)
}
\arguments{
\item{list_of_COSICC_DE_core_outputs}{list of outputs from COSICC_DE_core, one list element per lineage}

\item{names_lineage_trajectories}{names of the lineage trajectories associated with the output in
list_of_COSICC_DE_core_outputs}

\item{thresh_p_adj}{threshold for the adjusted p-value for selecting a gene as DE for a lineage trajectory}
}
\value{
input list list_of_COSICC_DE_core_outputs, barplots of numbers of DE genes for each lineage for static DE,
(p_static) and dynamic DE (p_dynamic), list of volcano plots for static DE (volcano_plot_list_static)
and dynamic DE (volcano_plot_list_dynamic)
}
\description{
combine_lineages_COSICC_DE
}
\details{
This function combines and plots outputs from COSICC_DE_core across differential lineage trajectories.
}
