% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions_kinetics_DE.R
\name{compute_and_plot_pseudotime}
\alias{compute_and_plot_pseudotime}
\title{compute_and_plot_pseudotime}
\usage{
compute_and_plot_pseudotime(sce)
}
\arguments{
\item{sce}{An object of type SingleCellExperiment including cells for one lineage, the colData of each of the SingleCellExperiments
in the list needs to include the slot 'stage' for embryonic stage.}
}
\description{
compute_and_plot_pseudotime
}
\details{
The function compute_and_plot_pseudotime computes diffusion maps, and identifies the diffusion component
most correlated with embryonic stage. The function also prints plots of the first two diffusion components,
}
