% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions_DA_testing_general.R
\name{COSICC_DA_lineage}
\alias{COSICC_DA_lineage}
\title{COSICC_DA_lineage}
\usage{
COSICC_DA_lineage(sce_case, sce_control, lineage_scores, alpha = 0.1)
}
\arguments{
\item{sce_case}{SingleCellExperiment for the case data set,
colData needs to include the slots 'marked' (TRUE/FALSE) for the presence
of a perturbation (CRISPR, drug, disease etc.) or fluorescent marking,
and 'cell_type'}

\item{sce_control}{SingleCellExperiment for the control data set,
colData needs to include the slots 'marked' (TRUE/FALSE) for the presence
of a perturbation (CRISPR, drug, disease etc.)}

\item{lineage_scores}{data frame whose first column is the cell-id, and whose remaining columns
are scores reflecting the probability that a cell forms part of the lineage, e.g.
Waddington-OT cell fate matrices.}

\item{alpha}{significance level, default is 0.1}
}
\value{
data.frame with the following columns: lineage, odds_ratio-odds ratio of the percentage of
marked positive cells among the target chimera cells belonging to the lineage versus
the respective ratio for the control experiment,
p_values, sig- whether a lineage is significantly enriched, depleted,
or neither significantly enriched nor depleted
}
\description{
COSICC_DA_lineage
}
\details{
Computes differential cell fate probability for perturbation experiments
relative to control experiments.
}
