% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions_kinetics_DE.R
\name{COSICC_DE}
\alias{COSICC_DE}
\title{COSICC_DE}
\usage{
COSICC_DE(
  list_sce_case,
  list_sce_control,
  thresh_p_adj = 0.1,
  list_lineage_genes
)
}
\arguments{
\item{list_sce_case}{A list of objects of class SingleCellExperiment for the perturbation data set
including cells for one lineage trajectory, containing one such SingleCellExperiment object for each lineage.
The names of the list are the names of the lineages.
The colData for each of the SingleCellExperiment objects needs to include columns labelled
cell in cancer etc.), 'lineage' and 'pt' (pseudotime, double, a measure of progression along the
lineage trajectory.}

\item{list_sce_control}{Like list_sce_case, for the control instead of the perturbation data set}

\item{thresh_p_adj}{double between 0 and 0.1, significance threshold for the adjusted p-values,
for selecting a gene as DE for a lineage trajectory}

\item{list_lineage_genes}{list of genes to be tested (ordered in the same way as list_sce_case and list_sce_control),
externally determined genes that are important to the progression along the lineage trajectories}
}
\value{
see combine_lineages_COSICC_DE
}
\description{
COSICC_DE

COSICC_DE integrates the functions COSICC_DE_core and combine_lineages_COSICC_DE.
}
