% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions_kinetics_DE.R
\name{COSICC_DE_core}
\alias{COSICC_DE_core}
\title{COSICC_DE_core}
\usage{
COSICC_DE_core(sce_case, sce_control, lineage_genes, N = 30)
}
\arguments{
\item{sce_case}{Object of class SingleCellExperiment for the control data set
including cells for one lineage trajectory. The colData needs to include the
following column: pt: observed time or pseudotime; marked: whether the
cell has a fluorescent or other marker (TRUE) or not (FALSE)}

\item{lineage_genes}{genes to be tested, externally determined genes that are important
to the progression along the lineage trajectory}

\item{N}{number of times the data set is split into two halves, where one half
is used for the identification of predictors using lasso, and the other half is
used for the computation of p-values. This avoid randomness in the p-values.}

\item{effect_size}{matrix with columns static and dynamic, one row for each gene.
The first column lists the log-fold-changes for static differential expression,
the second the effect sizes for the dynamic case.}

\item{p_val_adj_static}{adjusted p-values for static DE, based on correction
across all genes for the given lineage}

\item{p_val_adj_dynamic}{adjusted p-values for dynamic DE, based on correction
across all genes for the given lineage}
}
\value{
A list containing the following:
}
\description{
COSICC_DE_core
}
\details{
COSICC_DE_core uses lasso regression to determine whether a gene
is statically or dynamically dynamically expressed (DE) along a lineage trajectory.
p-values are computed following Meinshausen et al. (2009) JASA. By static differential
expresssion we mean that the gene is DE between marked and
unmarked cells, independently of the temporal or pseudotemporal state of the
specific cell. Dynamic DE, by contrast, refers to genes whose expression changes as a function of
(pseudo)time in different ways between marked and unmarked cells.
}
