% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions_kinetics_DE.R
\name{find_dynamic_genes}
\alias{find_dynamic_genes}
\title{find_dynamic_genes}
\usage{
find_dynamic_genes(sce, excluded_genes = NULL, FDR_across_time_points = 10^-10)
}
\arguments{
\item{sce}{An object of class SingleCellExperiment including cells for one lineage. The colData
needs to include columns labelled 'stage' and 'batch'.}

\item{excluded_genes}{Genes to be excluded a priory from the selection of dynamic genes,
by default NULL (none).}

\item{FDR_across_time_points}{see above}
}
\value{
An object of type SingleCellExperiment subset to the dynamic genes identified, and otherwise
identical to the input sce.
}
\description{
find_dynamic_genes
}
\details{
Identification of genes that vary significantly across stages (FDR < max_FDR_across_time_points)
and for which there is no strong evidence of variation within batches of the same stage for a reference data set.
}
