% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{simulate_DE_dynamic}
\alias{simulate_DE_dynamic}
\title{simulate_DE_dynamic}
\usage{
simulate_DE_dynamic(
  n,
  seeds = NULL,
  de.prob_case = 0.35,
  de.prob_control = 0.2,
  n_genes = 4,
  plot = TRUE
)
}
\arguments{
\item{n}{integer, number of lineages to be simulated}

\item{de.prob_case}{probability of differential expression for marked cells
compared to unmarked cells for the case data set}

\item{de.prob_control}{probability of differential expression for marked cells
compared to unmarked cells for the control data set}

\item{n_genes}{number of genes to be simulated}

\item{plot}{whether to produce gene expression plots for simulated genes}

\item{list_sce_case}{list of simulated perturbation data sets}

\item{list_sce_control}{list of simulated control data sets}
}
\value{
A nested list consisting of the following three lists

A list containing the following two lists
}
\description{
Simulation of dynamic differential expression, i.e differences between marked and
unmarked cells in terms of changes in expression along a lineage trajectory.
}
