% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sim_data.R
\name{simulate_sce_kinetics}
\alias{simulate_sce_kinetics}
\title{simulate_sce_kinetics: Simulate SingleCellExperiment objects with multiple 
lineaage trajectories and delay along some of the trajectories for perturbed cells}
\usage{
simulate_sce_kinetics(
  batchCells = 5000,
  nGenes = 10,
  de.prob = 0.5,
  nPaths = 5,
  seed = NULL,
  control_deplete_n = 10,
  case_deplete_n = "all",
  depleted_paths = c("Path1", "Path2", "Path3")
)
}
\arguments{
\item{batchCells}{Integer. Number of cells to simulate. Default is 500.}

\item{nGenes}{Integer. Number of genes to simulate. Default is 10.}

\item{de.prob}{Numeric. Probability of a gene being differentially expressed
along the lineage trajectory. 
Default is 0.5.}

\item{nPaths}{Integer. Number of trajectories (paths) to simulate. Default is 5.}

\item{seed}{Integer or NULL. Random seed for reproducibility. Default is NULL.}

\item{control_deplete_n}{Integer. Number of tdTomato+ high-pseudotime 
cells to deplete per lineage trajectory for the control data set. Default is 40.}

\item{case_deplete_n}{Integer. Number of tdTomato+ high-pseudotime cells 
to deplete per depleted trajectory for the case data set, e.g. the knockout chimera. Default is 80.}

\item{depleted_paths}{Character vector. Names of trajectories to deplete. 
Default is c("Path1", "Path2", "Path3").}
}
\value{
A named list containing two `SingleCellExperiment` objects:
\describe{
  \item{control}{SCE object representing the control condition. The \code{colData} contains:
    \itemize{
      \item \code{Cell}: Cell identifier
      \item \code{Path}: Lineage rajectory assignment (e.g., "Path1", "Path2", ...)
      \item \code{pt}: Pseudotime value for each cell
      \item \code{tdTomato}: Logical, whether the cell is tdTomato positive
      \item \code{nGene, nUMI, log10_total_counts, ...}: Additional QC columns from splatter
    }
  }
  \item{case}{SCE object representing the case condition, with additional depletion of tdTomato+ cells in selected trajectories. The \code{colData} columns are as above.}
}
}
\description{
Simulates control and case `SingleCellExperiment` objects using the 
splatter package's trajectory simulation.
Five lineage trajectories are simulated; for three of them, some tdTomato+ 
cells with higher pseudotime are depleted to simulate developmental delay, while for two, 
no such depletion is simulated.
The function allows for reproducibility via a user-specified random seed.
}
