% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sim_data.R
\name{simulate_seurat}
\alias{simulate_seurat}
\title{simulate_seurat: simulate a Seurat object as input data to 
COSICC_DA_group for testing and illustration}
\usage{
simulate_seurat(nGenes = 10, batchCells = 5000, nGroups = 10, seed = NULL)
}
\arguments{
\item{nGenes}{Integer. Number of genes to simulate. Default is 10.}

\item{batchCells}{Integer. Number of cells to simulate. Default is 5000.}

\item{nGroups}{Integer. Number of distinct cell types (groups). Default is 10.}

\item{seed}{Optional integer. Random seed for reproducibility. Default is `NULL`.}
}
\value{
A named list containing two `Seurat` objects:
\describe{
  \item{case}{Seurat object representing the case condition, including simulated depletion of tdTomato-positive cells in some groups.}
  \item{control}{Seurat object representing the control condition with randomly assigned tdTomato-positive cells.}
}
}
\description{
Generates simulated control and case `Seurat` objects using the
`splatter` package. A binary column `tdTomato` is added to each cell to simulate a
feature of interest, e.g. fluorescent marking of the cells.
}
\examples{
  sim_data <- simulate_seurat(seed = 123)
  seurat_case <- sim_data$case
  seurat_control <- sim_data$control
}
