% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.R
\name{test_FDR}
\alias{test_FDR}
\title{test_FDR}
\usage{
test_FDR(n, m)
}
\arguments{
\item{n}{number of lineage trajectories to be simulated for the test data set.}

\item{m}{number of genes to be simulated and tested for each lineage trajectory}
}
\description{
test_FDR
}
\details{
This function simulates data with no differential expression to test FDR control
for COSICC_DE.
}
