% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions_DA_testing_general.R
\name{COSICC_DA_group}
\alias{COSICC_DA_group}
\title{COSICC_DA_group}
\usage{
COSICC_DA_group(
  sce_case,
  sce_control,
  alpha = 0.1,
  thresh_marked_control = 5,
  thresh_unmarked_case = 5
)
}
\arguments{
\item{sce_case}{SingleCellExperiment for the case data set,
colData needs to include the slots 'marked' (TRUE/FALSE) for the presence
of a perturbation (CRISPR, drug, disease etc.) or fluorescent marking,
and 'cell_type'}

\item{sce_control}{SingleCellExperiment for the control data set,
colData needs to include the slots 'marked' (TRUE/FALSE) for the presence
of a perturbation (CRISPR, drug, disease etc.)}

\item{alpha}{significance level, default is 10% FDR}

\item{thresh_marked_control}{minimum number of marked cells per cell type in the control data set}

\item{thresh_unmarked_case}{minimum number of unmarked cells per cell type in the case data set}
}
\value{
data.frame with the following columns: cell_type, odds_ratio-odds ratio of the percentage of
marked cells among the cells in the case data versus the respective ratio for the control experiment,
p_values, sig- whether a cell_type is significantly enriched, depleted, or not enriched
}
\description{
COSICC_DA_group performs differential abundance testing
for perturbation experiments relative to control experiments.
}
